pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-ai-agent"
version = "0.1.0"
description = "Telegram Job Watcher (LLM-Driven) — Python + Telethon + Postgres"
readme = "README.md"
requires-python = ">=3.10,<3.14"
authors = [{ name = "You" }]
dependencies = [
  "telethon>=1.35.0",
  "SQLAlchemy>=2.0.25",
  "asyncpg>=0.29.0",
  "pydantic>=2.8.0",
  "tenacity>=8.2.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "openai>=1.40.0",
  "pytz>=2023.3",
  "greenlet>=2.0.0"
]

[tool.setuptools.packages.find]
include = ["app*"]


